# docs/enterprise-digital-twin/contracts/openapi/enterprise-digital-twin.openapi.yaml
openapi: 3.1.0
info:
  title: Enterprise Digital Twin API
  version: 1.0.0
  summary: Permission-aware graph, simulation, governed-action, and synthetic physical-asset API.
  description: Tenant context is always derived from the authenticated principal and a server-issued context handle. Tenant identifiers supplied in payloads are consistency assertions and never select authorization scope. Raw X-Tenant-ID is rejected.
...
servers:
  - url: https://api.enterprise-digital-twin.example
security:
  - oidc: [edt.read]
    contextHeader: []
  - oidc: [edt.read]
    contextCookie: []
...
paths:
  /v1/twin/simulation/snapshots:
    post:
      tags: [Simulation]
      operationId: createDecisionSimulationSnapshot
      description: Seals an immutable authorized graph-backed aggregate-variable snapshot. Tenant identity is server-derived.
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '../schemas/decision-intelligence.schema.json#/$defs/SnapshotCreate'}}}}
      responses:
        '201': {description: Immutable decision snapshot with canonical digest., headers: {ETag: {schema: {type: string}}}, content: {application/json: {schema: {type: object}}}}
        default: {$ref: '#/components/responses/Problem'}
  /v1/twin/simulation/snapshots/{snapshot_id}:
    get:
      tags: [Simulation]
      operationId: getDecisionSimulationSnapshot
      parameters: [{name: snapshot_id, in: path, required: true, schema: {type: string, format: uuid}}]
      responses:
        '200': {description: Tenant-scoped immutable decision snapshot., headers: {ETag: {schema: {type: string}}}, content: {application/json: {schema: {type: object}}}}
        default: {$ref: '#/components/responses/Problem'}
  /v1/twin/simulation/scenarios:
    post:
      tags: [Simulation]
      operationId: createDecisionScenario
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '../schemas/decision-intelligence.schema.json#/$defs/ScenarioCreate'}}}}
      responses:
        '201': {description: Draft root branch and confirmation digest., headers: {ETag: {schema: {type: string}}}, content: {application/json: {schema: {type: object}}}}
        default: {$ref: '#/components/responses/Problem'}
  /v1/twin/simulation/scenarios/{scenario_id}/branches:
    get:
      tags: [Simulation]
      operationId: listDecisionScenarioBranches
      parameters: [{name: scenario_id, in: path, required: true, schema: {type: string, format: uuid}}]
      responses:
        '200': {description: Up to 100 branches for the tenant scenario., content: {application/json: {schema: {type: object}}}}
        default: {$ref: '#/components/responses/Problem'}
    post:
      tags: [Simulation]
      operationId: createDecisionScenarioBranch
      parameters:
        - {name: scenario_id, in: path, required: true, schema: {type: string, format: uuid}}
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody: {required: true, content: {application/json: {schema: {$ref: '../schemas/decision-intelligence.schema.json#/$defs/BranchCreate'}}}}
      responses:
        '201': {description: Draft child branch bound to a confirmed parent., headers: {ETag: {schema: {type: string}}}, content: {application/json: {schema: {type: object}}}}
        default: {$ref: '#/components/responses/Problem'}
  /v1/twin/simulation/scenarios/{scenario_id}/branches/{branch_id}/confirm:
    post:
      tags: [Simulation]
      operationId: confirmDecisionScenarioBranch
      parameters:
        - {name: scenario_id, in: path, required: true, schema: {type: string, format: uuid}}
        - {name: branch_id, in: path, required: true, schema: {type: string, format: uuid}}
        - {$ref: '#/components/parameters/IdempotencyKey'}
        - {$ref: '#/components/parameters/IfMatch'}
      requestBody: {required: true, content: {application/json: {schema: {$ref: '../schemas/decision-intelligence.schema.json#/$defs/ScenarioConfirmation'}}}}
      responses:
        '201': {description: Immutable confirmed branch., headers: {ETag: {schema: {type: string}}}, content: {application/json: {schema: {type: object}}}}
        default: {$ref: '#/components/responses/Problem'}
  /v1/twin/simulation/runs:
    post:
      tags: [Simulation]
      operationId: runDecisionSimulation
      description: Executes confirmed changes and deterministic relationship propagation. It never invokes prediction or an LLM.
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '../schemas/decision-intelligence.schema.json#/$defs/SimulationRunCreate'}}}}
      responses:
        '201': {description: Durable simulation result and baseline comparison., content: {application/json: {schema: {type: object}}}}
        default: {$ref: '#/components/responses/Problem'}
  /v1/twin/simulation/runs/{simulation_id}:
    get:
      tags: [Simulation]
      operationId: getDecisionSimulation
      parameters: [{name: simulation_id, in: path, required: true, schema: {type: string, format: uuid}}]
      responses:
        '200': {description: Tenant-scoped durable simulation run., content: {application/json: {schema: {type: object}}}}
        default: {$ref: '#/components/responses/Problem'}
  /v1/twin/simulation/comparisons:
    post:
      tags: [Simulation]
      operationId: compareDecisionSimulationBranches
      requestBody: {required: true, content: {application/json: {schema: {$ref: '../schemas/decision-intelligence.schema.json#/$defs/SimulationComparisonCreate'}}}}
      responses:
        '200': {description: Baseline comparisons for two to ten distinct branch runs., content: {application/json: {schema: {type: object}}}}
        default: {$ref: '#/components/responses/Problem'}
...
  /v1/scenarios:
    post:
      tags: [Simulation]
      operationId: createScenario
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/ScenarioDraft'}}}
      responses:
        '201':
          description: Scenario draft compiled against a sealed snapshot; no simulation can run until digest confirmation.
          headers: {ETag: {description: Strong scenario-draft version used by If-Match on confirmation., schema: {type: string}}}
          content: {application/json: {schema: {$ref: '#/components/schemas/ScenarioDraftResource'}}}
        default: {$ref: '#/components/responses/Problem'}
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://identity.enterprise-digital-twin.example/.well-known/openid-configuration
    contextHeader:
      type: apiKey
      in: header
      name: X-EDT-Context
      description: Opaque short-lived audience-bound server-issued membership context handle. Raw tenant identifiers are rejected.
    contextCookie:
      type: apiKey
      in: cookie
      name: EDT-Context
      description: Browser-only HTTP-only Secure SameSite opaque context handle; CSRF controls apply to commands.
  parameters:
...
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 16, maxLength: 128}}
    IfMatch: {name: If-Match, in: header, required: true, schema: {type: string, minLength: 3, maxLength: 200}}
...
  responses:
    Problem:
      description: RFC 9457-compatible error with no unauthorized resource existence disclosure.
      content: {application/problem+json: {schema: {$ref: '../schemas/canonical.schema.json#/$defs/Problem'}}}
  schemas:
...
    ScenarioDraft:
      type: object
      additionalProperties: false
      required: [name, target_date, snapshot_id, expected_snapshot_hash, interventions, seed, sample_count]
      properties:
        name: {type: string, minLength: 1, maxLength: 200}
        target_date: {type: [string, 'null'], format: date}
        snapshot_id: {type: string, format: uuid}
        expected_snapshot_hash: {type: string, pattern: '^[a-f0-9]{64}$'}
        interventions: {type: array, minItems: 1, maxItems: 50, items: {$ref: '../schemas/scenario.schema.json#/$defs/Intervention'}}
        seed: {type: string, pattern: '^(0|[1-9][0-9]{0,19})$', description: Unsigned 64-bit decimal string; the runtime enforces the uint64 maximum.}
        sample_count: {const: 50000}
    ScenarioDraftResource:
      type: object
      additionalProperties: false
      required: [scenario_id, name, snapshot_id, scenario_digest, status, compiled_scenario]
      properties:
        scenario_id: {type: string, format: uuid}
        name: {type: string}
        snapshot_id: {type: string, format: uuid}
        scenario_digest: {type: string, pattern: '^[a-f0-9]{64}$'}
        status: {const: draft}
        compiled_scenario: {$ref: '#/components/schemas/CompiledScenarioDraft'}
    CompiledScenarioDraft:
      type: object
      additionalProperties: false
      required: [snapshot_id, snapshot_hash, name, target_date, model_version, calendar_version, compiler_version, seed, sample_count, interventions, assumptions, scenario_digest]
      properties:
        snapshot_id: {type: string, format: uuid}
        snapshot_hash: {type: string, pattern: '^[a-f0-9]{64}$'}
        name: {type: string, minLength: 1, maxLength: 200}
        target_date: {type: [string, 'null'], format: date}
        model_version: {const: pert-monte-carlo/1.0.0}
        calendar_version: {type: string, minLength: 1, maxLength: 100}
        compiler_version: {type: string, minLength: 1, maxLength: 100}
        seed: {type: string, pattern: '^(0|[1-9][0-9]{0,19})$'}
        sample_count: {const: 50000}
        interventions: {type: array, minItems: 1, maxItems: 50, items: {$ref: '../schemas/scenario.schema.json#/$defs/Intervention'}}
        assumptions: {type: array, items: {type: object, additionalProperties: false, required: [assumption_id, statement, status, source], properties: {assumption_id: {type: string, format: uuid}, statement: {type: string}, status: {enum: [confirmed, inferred_from_evidence, missing, contested]}, source: {enum: [user, evidence, engine_default, confirmed_imputation]}, evidence_ids: {type: array, uniqueItems: true, items: {type: string, format: uuid}}}}}
        scenario_digest: {type: string, pattern: '^[a-f0-9]{64}$'}
...